requirements.txt
tensorflow==2.7.0
opencv-python==4.5.5.62
tqdm==4.62.3
dynaconf==3.1.7
Pillow==9.0.0
streamlit==1.3.1
pdoc3==0.10.0